Uploading version 3.0.0
git-svn-id: https://plugins.svn.wordpress.org/nelio-ab-testing/trunk@955819 b8457f37-d9ea-0310-8a92-e5e31aec5664

+++ includes/models/alternatives/post-alternative-experiment.php
@@ -87,24 +87,18 @@ if( !class_exists( 'NelioABPostAlternativeExperiment' ) ) {
 			$this->set_winning_alternative( $winning_alt );
 		}
 
-		public function create_title_alternative( $name ) {
-			$alts         = $this->get_alternatives();
-			$fake_post_id = -1;
-			foreach ( $alts as $aux )
-				if ( $aux->get_value() <= $fake_post_id )
-					$fake_post_id = $aux->get_value() - 1;
-
-			$alternative = new NelioABAlternative();
-			$alternative->set_name( $name );
-			$alternative->set_value( $fake_post_id );
-			$this->add_local_alternative( $alternative );
-		}
-
 		public function create_empty_alternative( $name, $post_type ) {
-			if ( $post_type == NelioABExperiment::PAGE_ALT_EXP )
-				$post_type = 'page';
-			else
-				$post_type = 'post';
+			switch ( $post_type ) {
+				case NelioABExperiment::PAGE_ALT_EXP:
+					$post_type = 'page';
+					break;
+				case NelioABExperiment::POST_ALT_EXP:
+					$post_type = 'post';
+					break;
+				default:
+					return false;
+			}
+
 			$post = array(
 				'post_type'    => $post_type,
 				'post_title'   => $name,
@@ -114,18 +108,25 @@ if( !class_exists( 'NelioABPostAlternativeExperiment' ) ) {
 				'post_name'    => 'nelioab_' . rand(1, 10),
 			);
 
+			// Retrieve original post
+			$ori_post = get_post( $this->get_originals_id(), ARRAY_A );
+			if ( $ori_post )
+				$post['post_author'] = $ori_post['post_author'];
+
 			$post_id = wp_insert_post( $post, true );
+			if ( is_wp_error( $post_id ) )
+				return false;
+
 			$aux = get_post( $post_id, ARRAY_A );
+
+			// Update the post_name
 			$aux['post_name'] = 'nelioab_' . $post_id;
 			wp_update_post( $aux );
 
+			// Prepare custom metadata
 			add_post_meta( $post_id, '_is_nelioab_alternative', 'true' );
 
-			$alt = new NelioABAlternative();
-			$alt->set_name( $name );
-			$alt->set_value( $post_id );
-
-			$this->add_local_alternative( $alt );
+			return $post_id;
 		}
 
 		public function create_alternative_copying_content( $name, $src_post_id ) {
@@ -135,10 +136,11 @@ if( !class_exists( 'NelioABPostAlternativeExperiment' ) ) {
 			// Retrieve original post
 			$src_post = get_post( $src_post_id, ARRAY_A );
 			if ( !$src_post )
-				return;
+				return false;
 
 			// Create new empty post
 			$post_data = array(
+				'post_author'  => $src_post['post_author'],
 				'post_type'    => $src_post['post_type'],
 				'post_title'   => $src_post['post_title'],
 				'post_content' => $src_post['post_content'],
@@ -148,11 +150,8 @@ if( !class_exists( 'NelioABPostAlternativeExperiment' ) ) {
 			);
 			$new_post_id = wp_insert_post( $post_data, true );
 			if ( is_wp_error( $new_post_id ) )
-				return;
+				return false;
 			$new_post = get_post( $new_post_id, ARRAY_A );
-
-			// Override all information
-			NelioABWpHelper::override( $new_post_id, $src_post_id );
 
 			// Update the post_name
 			$new_post['post_name'] = 'nelioab_' . $new_post_id;
@@ -160,13 +159,11 @@ if( !class_exists( 'NelioABPostAlternativeExperiment' ) ) {
 
 			// Prepare custom metadata
 			add_post_meta( $new_post_id, '_is_nelioab_alternative', 'true' );
+
+			// Override all information
 			NelioABWpHelper::override( $new_post_id, $src_post_id );
 
-			$alt = new NelioABAlternative();
-			$alt->set_name( $name );
-			$alt->set_value( $new_post_id );
-
-			$this->add_local_alternative( $alt );
+			return $new_post_id;
 		}
 
 		protected function determine_proper_status() {
@@ -187,12 +184,22 @@ if( !class_exists( 'NelioABPostAlternativeExperiment' ) ) {
 			return NelioABExperimentStatus::READY;
 		}
 
+		public function add_local_alternative( $alt ) {
+			if ( $this->get_type() == NelioABExperiment::TITLE_ALT_EXP ) {
+				$fake_post_id = -1;
+				foreach ( $this->get_alternatives() as $aux )
+					if ( $aux->get_value() <= $fake_post_id )
+						$fake_post_id = $aux->get_value() - 1;
+				$alt->set_value( $fake_post_id );
+			}
+			parent::add_local_alternative( $alt );
+		}
+
 		public function save() {
 			require_once( NELIOAB_MODELS_DIR . '/account-settings.php' );
 
 			// 1. UPDATE OR CREATE THE EXPERIMENT
 			// -------------------------------------------------------------------------
-
 			$url = '';
 			if ( $this->get_id() < 0 ) {
 				$url = sprintf(
@@ -271,6 +278,23 @@ if( !class_exists( 'NelioABPostAlternativeExperiment' ) ) {
 				if ( $alt->was_removed() )
 					continue;
 
+				if ( $this->get_type() != NelioABExperiment::TITLE_ALT_EXP ) {
+					if ( $alt->is_based_on_a_post() ) {
+						$new_id = $this->create_alternative_copying_content( $alt->get_name(), $alt->get_base_post() );
+						if ( $new_id )
+							$alt->set_value( $new_id );
+						else
+							continue;
+					}
+					else {
+						$new_id = $this->create_empty_alternative( $alt->get_name(), $this->get_type() );
+						if ( $new_id )
+							$alt->set_value( $new_id );
+						else
+							continue;
+					}
+				}
+
 				$body = array(
 					'name'  => $alt->get_name(),
 					'value' => $alt->get_value(),
@@ -281,6 +305,7 @@ if( !class_exists( 'NelioABPostAlternativeExperiment' ) ) {
 					$result = NelioABBackend::remote_post(
 						sprintf( NELIOAB_BACKEND_URL . '/exp/post/%s/alternative', $exp_id ),
 						$body );
+					$alt->set_id( $result );
 				}
 				catch ( Exception $e ) {
 					// If I could not add an alternative... remove the associated page
@@ -298,7 +323,8 @@ if( !class_exists( 'NelioABPostAlternativeExperiment' ) ) {
 			foreach ( $all_alternatives as $alt ) {
 				if ( $alt->was_removed() ) {
 					// Delete permanently (skipping Trash)
-					wp_delete_post( $alt->get_value(), true );
+					if ( $alt->get_value() )
+						wp_delete_post( $alt->get_value(), true );
 				}
 			}
 
@@ -372,6 +398,18 @@ if( !class_exists( 'NelioABPostAlternativeExperiment' ) ) {
 
 			// If everything is OK, we can start it!
 
+			// ...but before, if it is a title experiment, we have to create the goal
+			if ( $this->get_type() == NelioABExperiment::TITLE_ALT_EXP ) {
+				require_once( NELIOAB_MODELS_DIR . '/goals/page-accessed-goal.php' );
+				$goal = new NelioABPageAccessedGoal( $this );
+				$page = new NelioABPageDescription( $this->get_originals_id() );
+				$page->set_indirect_navigations_enabled( true );
+				$goal->add_page( $page );
+				$this->add_goal( $goal );
+				$this->make_goals_persistent();
+			}
+
+			// And there we go!
 			$ori_post = get_post( $this->get_originals_id() );
 			if ( $ori_post && $this->get_type() != NelioABExperiment::TITLE_ALT_EXP ) {
 				foreach ( $this->get_alternatives() as $alt ) {
@@ -387,12 +425,10 @@ if( !class_exists( 'NelioABPostAlternativeExperiment' ) ) {
 					$this->get_id()
 				);
 			try {
-				$this->split_page_accessed_goal_if_any();
 				$result = NelioABBackend::remote_post( $url );
 				$this->set_status( NelioABExperimentStatus::RUNNING );
 			}
 			catch ( Exception $e ) {
-				$this->unsplit_page_accessed_goal_if_any();
 				throw $e;
 			}
 		}
@@ -452,4 +488,3 @@ if( !class_exists( 'NelioABPostAlternativeExperiment' ) ) {
 
 }
 
-?>
